requirements.txt
deep-translator
docx
email-validator
Flask
Flask>=3.1.2
flask-login
flask-sqlalchemy
fuzzywuzzy
google-generativeai
groq
gunicorn
huggingface-hub
pdf2docx
Pillow
psycopg2-binary
pydub
PyMuPDF
pytesseract
python-docx
python-Levenshtein
requests
requests>=2.32.5
SpeechRecognition
sqlalchemy
werkzeug
yt-dlp
yt-dlp>=2025.01.01
